Add broken UI fixes:
Before, when running near/at the bottom of the screen, the cursor was being
saved such that redrawing would cause the previously rendered content to
be pushed up, which was annoying to say the least. Now, 8 lines are added
(8 newlines, followed by moving cursor up 8 lines), which should ensure there
is at least 8 lines at any given time. This still might not be enough though,
especially for the item list on the day page, so I added a global ^L (form
feed) handler, which will clear the screen and move your cursor to the top
of the screen when pressed.